+++ CHANGELOG.md
@@ -2,3 +2,4 @@ Rev#001
 -------
 - The program now uses FFT to perform the cross-correlation
 - Output is saved as a binary file
+- Now the program needs `argv[1]` and `argv[2]` as input matrix files